pyproject.toml
[tool.poetry]
name = "sakuramedia-llm"
version = "0.1.0"
description = "Lightweight async ASR service for SakuraMedia"
authors = ["tinyping"]
readme = "README.md"
packages = [{ include = "app" }]

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.116.0"
uvicorn = { extras = ["standard"], version = "^0.35.0" }
peewee = "^3.17.9"
pydantic-settings = "^2.10.1"
python-multipart = "^0.0.20"
faster-whisper = "^1.1.1"
ctranslate2 = "^4.6.0"
loguru = "^0.7.3"
tqdm = "^4.67.1"
onnxruntime = "<=1.16.3"
numpy = "<2"

[tool.poetry.group.dev.dependencies]
pytest = "^8.4.1"
pytest-cov = "^6.2.1"
httpx = "^0.28.1"

[build-system]
requires = ["poetry-core>=1.8.0"]
build-backend = "poetry.core.masonry.api"

[tool.pytest.ini_options]
pythonpath = ["."]
testpaths = ["tests"]
